chore: rename graph.py for clarity
