+++ agent_main.py
@@ -1,7 +1,7 @@
 from langsmith import traceable
 from langgraph.graph.state import CompiledStateGraph
 
-from src.agents.graph import create_research_graph
+from src.agents.report_graph import create_research_graph
 from src.agents.utils.email_service import MailingService
 from src.agents.utils.output_utils import (
     convert_markdown_to_pdf,
